[ios, macos] Short-circuit redundant camera changes
Avoid canceling transitions (and triggering preexisting completion handlers) until we get a chance to ensure that a new transition really does have to begin. Consistently avoid mbgl transitions for redundant camera changes. Upon bailing, schedule the completion handler to run asynchronously on a delay equivalent to the requested animation duration.

MGLMapCamera now performs a float-precision equality check to avoid trivial differences. Added an equality operator to CameraOptions that similarly uses float-precision.

+++ platform/darwin/src/MGLMapCamera.mm
@@ -110,16 +110,17 @@
     }
     
     MGLMapCamera *otherCamera = other;
-    return (_centerCoordinate.latitude == otherCamera.centerCoordinate.latitude
-            && _centerCoordinate.longitude == otherCamera.centerCoordinate.longitude
-            && _altitude == otherCamera.altitude
-            && _pitch == otherCamera.pitch && _heading == otherCamera.heading);
+    return ((float)_centerCoordinate.latitude == (float)otherCamera.centerCoordinate.latitude
+            && (float)_centerCoordinate.longitude == (float)otherCamera.centerCoordinate.longitude
+            && (float)_altitude == (float)otherCamera.altitude
+            && (float)_pitch == (float)otherCamera.pitch
+            && (float)_heading == (float)otherCamera.heading);
 }
 
 - (NSUInteger)hash
 {
-    return (@(_centerCoordinate.latitude).hash + @(_centerCoordinate.longitude).hash
-            + @(_altitude).hash + @(_pitch).hash + @(_heading).hash);
+    return (@((float)_centerCoordinate.latitude).hash + @((float)_centerCoordinate.longitude).hash
+            + @((float)_altitude).hash + @((float)_pitch).hash + @((float)_heading).hash);
 }
 
 @end
